Keep it simple, by explicitly requiring modules.
Webpack was choking on requiring the index.js file, because
it was trying to use the fs module to iterate over files.

There's also no real reason to get fancy here, because hardly
any typing is saved, and even if you saved some chars here,
it's not like this is boilerplate the user has to type over
and over.

Webpack and browserify work through an AST provided by esprima,
so you have to explicitly specify the strings of the modules,
(with a few exceptions).